+++ Base Project/Base Project/src/main/java/com/example/baseproject/WithdrawController.java
@@ -56,14 +56,14 @@ public class WithdrawController  {
         String value = withdrawController.getValue(valueField);
         String getBalance = "SELECT balance from user_account where username='" + username + "'";
         ResultSet bal = null;
-        int intValue = 0;
+        double intValue = 0;
         try {
             Statement statement = connectDB.createStatement();
             bal = statement.executeQuery(getBalance);
             while (bal.next()) {
                 String balance = bal.getString("balance");
-                intValue = Integer.parseInt(value);
-                intValue = Integer.parseInt(String.valueOf(balance)) - intValue;
+                intValue = Double.parseDouble(value);
+                intValue = Double.parseDouble(String.valueOf(balance)) - intValue;
             }
         } catch (SQLException e) {
             e.printStackTrace();
@@ -71,7 +71,7 @@ public class WithdrawController  {
         if (intValue < 0) {invalid.setText("Insufficient Funds!!!");ok=0;
             }
         else {
-            value = Integer.toString(intValue);
+            value = Double.toString(intValue);
             try {
                 connection = DriverManager.getConnection("jdbc:mysql://localhost/usersdb", "root", "root");
                 psInsert = connection.prepareStatement("UPDATE user_account SET balance =? WHERE username=?");
